order ok

+++ Sspos/www/js/sql.js
@@ -3,7 +3,7 @@ sql = (function() {
 	var ret = {};
 	var db = "";
 	var timeList = [5000, 10000, 20000, 30000]
-    var countipc = 0, countisi = 0, countipt = 0;
+    var countipc = 0, countisi = 0, countipt = 0, countipmt = 0;
 	ret._allnid = {};
 	//打开数据库
 	ret.opendb = function(){
@@ -21,10 +21,13 @@ sql = (function() {
 	        tx.executeSql('create table if not exists product_category (nid text primary key,title text,start_time text,end_time text,pid INTEGER,image text)');
 	        //创建产品表
 	        tx.executeSql('DROP TABLE IF EXISTS product');
-	        tx.executeSql('create table if not exists product (sku text primary key,title text,price text,sold_out text,nid text,image text,imageurl text,time_period text,weight text,isreplace text)');
+	        tx.executeSql('create table if not exists product (sku text primary key,title text,price text,sold_out text,nid text,image text,imageurl text,time_period text,weight INTEGER,isreplace text)');
 	        //创建订单表
 	        tx.executeSql('DROP TABLE IF EXISTS orderinfo');
 	        tx.executeSql('create table if not exists orderinfo (orderid text primary key,payamount text,paystatus text,paytype text,discountamount text,productlist text,addtime text)');
+	        //支付信息表
+	        tx.executeSql('DROP TABLE IF EXISTS pmtinfo');
+	        tx.executeSql('create table if not exists pmtinfo (channel_code text,channel_name text,isflag INTEGER)');
 		},function(e){
 			console.log("创建数据表失败");
 		})
@@ -32,26 +35,31 @@ sql = (function() {
 	//插入门店信息
 	ret.insertStoreinfo = function(fun){
 		de.get_store(function(result){
-        	if(result.state != 1||result.data == ""){
-        		console.log("error: get storeinfo is null");
-        		fun(false);
-        	}else{
-        		var store = result.data;
-        		config.STORENAME= store.title;
-                config.address_detail=store.address_detail;
-        		db.transaction(function (tx) {
-        			tx.executeSql('delete from storeinfo'); //防止插入多条数据
-        			tx.executeSql('insert or replace into storeinfo (storeid,title,address,address_detail,pid,version,update_detime) values(?,?,?,?,?,?,?)',
-        				[store.store_id,store.title,store.address.name,store.address_detail,config.ID,config.VERSION,new Date().format('yyyy-MM-dd')]
-        				,function(tx,res){
-        					console.log("success: insert storeinfo is ok");
-        					fun(true);
-        			},function(e){
-        				console.log("error: insert storeinfo error");
-        				fun(false);
-        			})
-        		});
-        	}
+			if(result.hasOwnProperty("state")) {
+				if(result.state != 1||result.data == ""){
+	        		console.log("error: get storeinfo is null");
+	        		fun(false);
+	        	}else{
+	        		var store = result.data;
+	        		config.STORENAME= store.title;
+	                config.address_detail=store.address_detail;
+	        		db.transaction(function (tx) {
+	        			tx.executeSql('delete from storeinfo'); //防止插入多条数据
+	        			tx.executeSql('insert or replace into storeinfo (storeid,title,address,address_detail,pid,version,update_detime) values(?,?,?,?,?,?,?)',
+	        				[store.store_id,store.title,store.address.name,store.address_detail,config.ID,config.VERSION,new Date().format('yyyy-MM-dd')]
+	        				,function(tx,res){
+	        					console.log("success: insert storeinfo is ok");
+	        					fun(true);
+	        			},function(e){
+	        				console.log("error: insert storeinfo error");
+	        				fun(false);
+	        			})
+	        		});
+	        	}
+			}else{
+				console.log("error: get storeinfo error");
+	        	fun(false);
+			}
 	    },function(jqXHR, textStatus, errorThrown){
 	        console.log(jqXHR);
 	        console.log("error: get storeinfo cannot access");
@@ -61,27 +69,33 @@ sql = (function() {
 	//插入产品分类信息
 	ret.insertProduct_Category = function(fun){
 		de.get_product_category_nbs(function(result){
-			if(result.state !=1||result.data == ""){
-				console.log("error: get product_category is null");
-				fun(false);
-			}else{
-				var categorys = result.data;
-				db.transaction(function(tx){
-					tx.executeSql('delete from product_category');
-					for (var va in categorys) {
-						tx.executeSql('insert or replace into product_category (title,nid,start_time,end_time,pid,image,image2) values(?,?,?,?,?,?)',
-                                [categorys[va].name,categorys[va].tid,"00:00","23:59",categorys[va].weight,"images/zc.png"],
-							function(tx,res){
+			if(result.hasOwnProperty("state")) {
+				if(result.state !=1||result.data == ""){
+					console.log("error: get product_category is null");
+					fun(false);
+				}else{
+					var categorys = result.data;
+					db.transaction(function(tx){
+						tx.executeSql('delete from product_category');
+						for (var va in categorys) {
+							tx.executeSql('insert or replace into product_category (title,nid,start_time,end_time,pid,image) values(?,?,?,?,?,?)',
+	                                [categorys[va].name,categorys[va].tid,"00:00","23:59",categorys[va].weight,"images/zc.png"],
+								function(tx,res){
 
-						},function(e){
-							console.log("error: insert product_category is error");
-							fun(false);
-						})
-					}
-					console.log("success: insert product_category is ok");
-					fun(true);
-				});
+							},function(e){
+								console.log("error: insert product_category is error");
+								fun(false);
+							})
+						}
+						console.log("success: insert product_category is ok");
+						fun(true);
+					});
+				}
+			} else {
+				console.log("error: get product_category error");
+				fun(false);
 			}
+			
 		},function(jqXHR, textStatus, errorThrown){
 			console.log(jqXHR);
 	        console.log("error: get product_category cannot access");
@@ -91,62 +105,67 @@ sql = (function() {
 	//插入产品信息
 	ret.insertproducts = function(fun){
 		de.get_products(function(result){
-			if(result.state !=1||result.data == ""){
-				console.log("error: get products is null");
-				fun(false);
-			}else{
-				var products = result.data;
-				db.transaction(function(tx){
-					tx.executeSql('delete from product');
-					for(var i=0;i<products.length;i++){
-						var product = products[i];
-						var nid = "";
-						var timelist = new Array();
-						for (var j = 0; j < product.time_period.length; j++) {   
-                           var _start = product.time_period[j].times.start;
-                           var _end = product.time_period[j].times.end;
-                           if(_start.length<5){
-                            _start = "0" + _start;
-                           }
-                           if(_end.length<5){
-                            _end = "0" + _end;
-                           }
-                           timelist.push(_start);
-                           timelist.push(_end);
-                        }
-                        timelist.sort();
-                        var my_image="";
-                        if(product.sku.indexOf(".")>-1||product.sku.indexOf("_")>-1){
-                            my_image = product.image_small
-                        }else{
-                            my_image = product.image_big;
-                        }
-                        if (product.category_nbs.length <= 0) {
-                            continue;
-                        }else{
-                        	for (var k = 0; k < product.category_nbs.length; k++) {
-                                nid = nid + product.category_nbs[k].tid + ",";
-                                if(!ret._allnid.hasOwnProperty(product.category_nbs[k].tid)){
-                                    ret._allnid[product.category_nbs[k].tid] = [];
-                                    ret._allnid[product.category_nbs[k].tid].push(timelist[0]);
-                                    ret._allnid[product.category_nbs[k].tid].push(timelist[timelist.length-1]);
-                                }else{
-                                    ret._allnid[product.category_nbs[k].tid].push(timelist[0]);
-                                    ret._allnid[product.category_nbs[k].tid].push(timelist[timelist.length-1]);
-                                }
-                            }
-                        }
-                        tx.executeSql('insert or replace into product (sku,title,price,sold_out,nid,image,imageurl,time_period,weight,isreplace) values(?,?,?,?,?,?,?,?,?,?)',
-                                [product.sku, product.title, product.price, product.sold_out, nid, ret.getThumbNameByUrl(my_image), my_image,JSON.stringify(product.time_period),product.weight,true],
-                                function(tx,res){
+			if(result.hasOwnProperty("state")) {
+				if(result.state !=1||result.data == ""){
+					console.log("error: get products is null");
+					fun(false);
+				}else{
+					var products = result.data;
+					db.transaction(function(tx){
+						tx.executeSql('delete from product');
+						for(var i=0;i<products.length;i++){
+							var product = products[i];
+							var nid = "";
+							var timelist = new Array();
+							for (var j = 0; j < product.time_period.length; j++) {   
+	                           var _start = product.time_period[j].times.start;
+	                           var _end = product.time_period[j].times.end;
+	                           if(_start.length<5){
+	                            _start = "0" + _start;
+	                           }
+	                           if(_end.length<5){
+	                            _end = "0" + _end;
+	                           }
+	                           timelist.push(_start);
+	                           timelist.push(_end);
+	                        }
+	                        timelist.sort();
+	                        var my_image="";
+	                        if(product.sku.indexOf(".")>-1||product.sku.indexOf("_")>-1){
+	                            my_image = product.image_small
+	                        }else{
+	                            my_image = product.image_big;
+	                        }
+	                        if (product.category_nbs.length <= 0) {
+	                            continue;
+	                        }else{
+	                        	for (var k = 0; k < product.category_nbs.length; k++) {
+	                                nid = nid + product.category_nbs[k].tid + ",";
+	                                if(!ret._allnid.hasOwnProperty(product.category_nbs[k].tid)){
+	                                    ret._allnid[product.category_nbs[k].tid] = [];
+	                                    ret._allnid[product.category_nbs[k].tid].push(timelist[0]);
+	                                    ret._allnid[product.category_nbs[k].tid].push(timelist[timelist.length-1]);
+	                                }else{
+	                                    ret._allnid[product.category_nbs[k].tid].push(timelist[0]);
+	                                    ret._allnid[product.category_nbs[k].tid].push(timelist[timelist.length-1]);
+	                                }
+	                            }
+	                        }
+	                        tx.executeSql('insert or replace into product (sku,title,price,sold_out,nid,image,imageurl,time_period,weight,isreplace) values(?,?,?,?,?,?,?,?,?,?)',
+	                                [product.sku, product.title, product.price, product.sold_out, nid, ret.getThumbNameByUrl(my_image), my_image,JSON.stringify(product.time_period),product.weight,true],
+	                                function(tx,res){
 
-                        },function(e){
-                        	console.log(e);
-                        })
-					}
-					console.log("success: insert products is ok");
-					fun(true);
-				})
+	                        },function(e){
+	                        	console.log(e);
+	                        })
+						}
+						console.log("success: insert products is ok");
+						fun(true);
+					})
+				}
+			} else {
+				console.log("error: get products error");
+				fun(false);
 			}
 		},function(jqXHR, textStatus, errorThrown){
 			console.log(jqXHR);
@@ -158,16 +177,52 @@ sql = (function() {
 	ret.insertOrderInfo = function(param,fun){
 		db.transaction(function(tx){
 			tx.executeSql('insert into orderinfo (orderid,payamount,paystatus,paytype,discountamount,productlist,addtime) values(?,?,?,?,?,?,?)',
-				[],function(tx,res){
-				console.log("success: insert orderinfo is");
-	       		fun(false);
+				[param.orderInfo.orderid,JSON.stringify(param.orderInfo.payamount),JSON.stringify(param.orderInfo.paystatus),param.orderInfo.paytype,JSON.stringify(param.orderInfo.dicountamount),
+				JSON.stringify(param.productList),param.orderInfo.booktime],function(tx,res){
+				fun(true);
 			},function(e){
-
+		        console.log("error: insert orderinfo is error");
+		        fun(false);
 			})
 		});
 	}
-	//转换url
-	ret.getThumbNameByUrl = function(url){
+	//插入支付信息表
+	ret.insertpmtinfo = function(param,fun) {
+		pmt.get_corporation_channel(param,function(result){
+			if(result.hasOwnProperty("channels")) {
+				db.transaction(function(tx) {
+					for(var va in result.channels) {
+						if(va == "alipay" || va == "wechat") {
+							for(var i=0; i<result.channels[va].length; i++) {
+								if(result.channels[va][i].channel_code == "create_direct_pay_by_user" || 
+									result.channels[va][i].channel_code == "NATIVE") {
+									var channel_code = result.channels[va][i].channel_code;
+									var channel_name = result.channels[va][i].channel_name;
+									var isflag = 1 ; //默认启用状态
+									tx.executeSql('insert or replace into pmtinfo (channel_code,channel_name,isflag) values(?,?,?)',
+										[channel_code,channel_name,isflag],function(tx,res) {
+
+									},function(e) {
+										console.log("error: insert pmtinfo error");
+										fun(false);
+									})
+								}
+							}
+						}
+					}
+				})
+			} else {
+				console.log("error: get pmtinfo error");
+				fun(false);
+			}
+		},function(jqXHR, textStatus, errorThrown){
+			console.log(jqXHR);
+	        console.log("error: get pmtinfo cannot access");
+	        fun(false);
+		})
+	}
+	//转换图片地址
+	ret.getThumbNameByUrl = function(url) {
 		var name = "";
 	    if (url && typeof (url) == "string") {
 	        var arr = url.split("/");
@@ -178,7 +233,7 @@ sql = (function() {
 	    return name;
 	}
 	//修改产品分类售卖时间
-	ret.updateCategoryTime = function(param,fun){
+	ret.updateCategoryTime = function(param,fun) {
 		db = ret.checkdb(db);
 		db.transaction(function (tx) {
 	        for(var anid in param){
@@ -193,7 +248,7 @@ sql = (function() {
 	    });
 	}
 	//修改posid
-	ret.updateStoreInfo_pid = function(param,fun){
+	ret.updateStoreInfo_pid = function(param,fun) {
 		db = ret.checkdb(db);
 		db.transaction(function (tx) {
 			tx.executeSql('update storeinfo set pid=? where storeid=?',[param.pid,param.storeid],function(tx,res){
@@ -206,7 +261,7 @@ sql = (function() {
 		})
 	}
 	//修改产品售罄信息
-	ret.updateProduct_sold_out = function(param,fun){
+	ret.updateProduct_sold_out = function(param,fun) {
 		db = ret.checkdb(db);
 		db.transaction(function (tx) {
 			for(var va in param){
@@ -222,8 +277,20 @@ sql = (function() {
 	        fun(true);	
 		})
 	}
+	ret.updateOrderInfo_paystatus = function(param,fun) {
+		db = ret.checkdb(db);
+		db.transaction(function (tx) {
+			tx.executeSql('update orderinfo set paystatus=? where orderid=?',[param.pid,param.merchant_no],function(tx,res){
+				console.log("success: update updateOrderInfo_paystatus success");
+	        	fun(true);
+			},function(e){
+				console.log("error: update updateOrderInfo_paystatus error");
+	        	fun(false);
+			})
+		})
+	}
 	//查询菜品分类
-	ret.queryCategory = function(fun){
+	ret.queryCategory = function(fun) {
 		db = ret.checkdb(db);
 		db.transaction(function (tx) {
 			tx.executeSql('SELECT * FROM product_category ORDER BY pid ASC', [], function(tx,res){
@@ -250,6 +317,7 @@ sql = (function() {
 			})
 		})
 	}
+	
 	//检查数据库是否打开
 	ret.checkdb = function(db){
 		if (db == undefined || db == null) {
@@ -306,6 +374,21 @@ sql = (function() {
                 });
             }
         });
+    }
+    //支付信息
+    ret.ipmt = function() {
+    	db = ret.checkdb(db);
+    	var _param = {
+    		sid : config.STOREID
+    	}
+    	ret.insertpmtinfo(_param,function(va) {
+    		if (countipmt > 3) {
+    			countipmt = 0;
+    		}if (!va) {
+    			setTimeout(ret.ipmt, timeList[countipt]);
+                countipmt++;
+    		}
+    	})
     }
 	return ret;
 })();
